requirements.txt
fastapi==0.109.0
telethon==1.33.1
python-dotenv==1.0.0
uvicorn==0.27.0
python-multipart==0.0.6
pydantic==2.5.3